fix: Entity Table 싱크

+++ src/main/kotlin/com/example/daitssuapi/domain/main/dto/response/UserResponse.kt
@@ -1,7 +1,7 @@
 package com.example.daitssuapi.domain.main.dto.response
 
 data class UserResponse(
-    val studentId: Int,
+    val studentId: Long,
     val name: String,
     val nickname: String,
     val departmentName: String,
